most of the components are moved into their own files
