+++ src/app.tsx
@@ -1,302 +1,24 @@
 import * as React from 'react'
-import { range, rangeFromTo } from 'lang/ranges';
+import { range } from 'lang/ranges';
 import { removeFromArray, shuffle, unique } from 'lang/arrays';
-import { noop, NoState } from 'lang/react';
-import StackPopped from 'stacks/stackPopped';
-import BringEventToConclusion from 'events/bringEventToConclusion';
-import PartyBattleGoals from 'battlegoals/partyBattleGoals';
-import AddCards, { AddCardsProps } from 'cards/addCards';
+import { PartyBattleGoals } from 'battlegoals/partyBattleGoals';
+import { AddCards } from 'cards/addCards';
 import { Draw } from 'cards/draw';
 import { CardStack } from 'cards/cards';
-import { OpenDialog } from 'app/app';
-import { RandomItemDesignProps, PersonalGoalProps, ItemUrl, ItemProps, RandomSideScenarioProps } from "cards/CardRenderProps";
-import { cardToDiv } from "cards/cards";
+import { RandomItemDesignProps, PersonalGoalProps, RandomSideScenarioProps } from 'cards/CardRenderProps';
+import { Pop } from 'stacks/Pop';
+import { Shop, ProsperityInput } from 'app/Shop';
+import { itemsAboveProsperity } from 'app/Propsperity';
+import { ImportExport } from 'app/ImportExport';
+import { RandomItemDesigns} from 'designs/RandomItemDesigns';
 
 const randomItemDesigns = new RandomItemDesignProps();
 const personalGoals = new PersonalGoalProps();
-
-const itemIdsByProsperityLevel: { [prosperityLevel: number]: Array<number> } = {
-  1: range(1, 14),
-  2: range(15, 7),
-  3: range(22, 7),
-  4: range(29, 7),
-  5: range(36, 7),
-  6: range(43, 7),
-  7: range(50, 7),
-  8: range(57, 7),
-  9: range(64, 7),
-};
-
-const itemUrls = function (): Array<ItemUrl> {
-  const itemCounts = [0, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 4, 4, 4, 2, 2, 2, 2, 2, 4, 2, 2, 2, 2, 2, 2, 4, 2, 2, 2, 2, 2, 2, 4, 2, 2, 2, 2, 2, 2, 4, 2, 2, 2, 2, 2, 2, 4, 2, 2, 2, 2, 2, 2, 4, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 1, 1, 1, 1, 2, 1, 1, 1, 2, 2, 2, 1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1];
-  const itemUrls = [{
-    maxItem: 20,
-    url: "https://lh3.googleusercontent.com/u/0/d/1P0bd7vtA_SVwC7Qm9dJ_YmZkQPNDLOCk=s3200-k-iv2",
-  }, {
-    maxItem: 40,
-    url: "https://lh3.googleusercontent.com/u/0/d/1uHYherEvc9bv3Jpl2TpA2DlRPgGT8l3z=s3200-k-iv2",
-  }, {
-    maxItem: 60,
-    url: "https://lh3.googleusercontent.com/u/0/d/172NPm8x9T8zPE2Vd672_10rZ3ieFHHOX=s3200-k-iv2",
-  }, {
-    maxItem: 90,
-    url: "https://lh3.googleusercontent.com/u/0/d/1XAabPK_Hs8gBXXJVpaBaSsAv6_NioEyu=s3200-k-iv2",
-  }, {
-    maxItem: 133,
-    url: "https://lh3.googleusercontent.com/u/0/d/1KW0TZOs7SDVl5frTM9y-ISR-4dZFIex4=s3200-k-iv2",
-  }, {
-    maxItem: 150,
-    url: "https://lh3.googleusercontent.com/u/0/d/1BHfEwqmC_dax5dV4RFFP5MlfJ76eS7KZ=s3200-k-iv2",
-  }];
-
-  let acc = 0;
-  let itemSheet = 0;
-  return itemCounts.map((num, i) => {
-    if (i >= 71 && i <= 95) {
-      // there are two random items but only one of them is "red", i.e. in this picture
-      num -= 1;
-    }
-    let numberInPicture = acc;
-    if (i > itemUrls[itemSheet].maxItem) {
-      itemSheet += 1;
-      numberInPicture = 0;
-      acc = num;
-    } else {
-      acc += num;
-    }
-    return new ItemUrl(itemUrls[itemSheet].url, numberInPicture, itemUrls[itemSheet].maxItem - (itemUrls[itemSheet - 1] && itemUrls[itemSheet - 1].maxItem || 0));
-  });
-}();
 
 const randomScenarios = {
   offset: 63,
   n: 9,
 };
-
-function itemToDiv(itemId: number) {
-  let item = itemUrls[itemId];
-  return cardToDiv(itemId, new ItemProps(item, itemId - item.numberInPicture));
-}
-
-interface RandomItemDesignsProps {
-  list: Array<number>
-}
-
-class RandomItemDesigns extends React.Component<RandomItemDesignsProps, NoState> {
-  render() {
-    if (this.props.list.length === 0) {
-      return null
-    } else {
-      return [
-        <h2 key="h2">Random Item Designs</h2>,
-        <div key="cards">
-          {this.props.list.map(i => itemToDiv(i))}
-        </div>
-      ]
-    }
-  }
-}
-
-interface PopProps extends OpenDialog, AddCardsProps, StackPopped {
-  name: string,
-  cards: CardStack
-}
-
-// Pop draws the _top_ card of the deck
-class Pop extends React.Component<PopProps, NoState> {
-  constructor(props: PopProps) {
-    super(props);
-    this.clicked = this.clicked.bind(this);
-  }
-
-  clicked() {
-    this.props.setDialog(
-      <BringEventToConclusion
-        name={this.props.name}
-        number={this.props.cards.stack[0]}
-        stackPopped={this.props.stackPopped}>
-        <AddCards onAddCards={this.props.onAddCards}/>
-      </BringEventToConclusion>
-    );
-  }
-
-  render() {
-    return <button type="button" onClick={this.clicked} disabled={this.props.cards.stack.length === 0}>{"Draw " + this.props.name + " Event"}</button>
-  }
-}
-
-interface ProsperityInputProps {
-  onIncreaseProsperity: () => void,
-  prosperity: number,
-}
-
-class ProsperityInput extends React.Component<ProsperityInputProps, NoState> {
-  constructor(props: ProsperityInputProps) {
-    super(props);
-    this.increaseProsperity = this.increaseProsperity.bind(this);
-  }
-
-  increaseProsperity() {
-    this.props.onIncreaseProsperity();
-  }
-
-  render() {
-    const prosperity = this.props.prosperity;
-    return <h2 key="h2">
-      Prosperity {prosperity}
-      <button disabled={(prosperity >= 9)} type="button" onClick={this.increaseProsperity}>+</button>
-    </h2>;
-  }
-}
-
-interface ShopProps {
-  prosperity: number,
-
-}
-
-enum ShopItemFilter {
-  All = 'all',
-}
-
-interface ShopState {
-  shopItemFilter: ShopItemFilter
-}
-
-class Shop extends React.Component<ShopProps, ShopState> {
-  static levelWithItems(level: number, items: Array<number>) {
-    return { level, items }
-  }
-
-  constructor(props: ShopProps) {
-    super(props);
-    this.handleShopItemFilterChange = this.handleShopItemFilterChange.bind(this);
-    this.itemsToDisplay = this.itemsToDisplay.bind(this);
-
-    this.state = {
-      shopItemFilter: ShopItemFilter.All
-    };
-  }
-
-  handleShopItemFilterChange(event: React.ChangeEvent<HTMLSelectElement>) {
-    this.setState({ shopItemFilter: event.target.value as ShopItemFilter }, noop);
-    event.preventDefault();
-  }
-
-  itemsToDisplay() {
-    const filterAsString = this.state.shopItemFilter;
-    switch (filterAsString) {
-      case 'all':
-        return rangeFromTo(1, this.props.prosperity + 1).map(level => Shop.levelWithItems(level, itemIdsByProsperityLevel[level]));
-      default:
-        const maybeProsperity = parseInt(filterAsString, 10);
-        const prosperityLevel = isNaN(maybeProsperity) ? 1 : maybeProsperity;
-        return [Shop.levelWithItems(prosperityLevel, itemIdsByProsperityLevel[prosperityLevel])];
-    }
-  }
-
-  render() {
-    const prosperity = this.props.prosperity;
-    return <div key="cards">
-      <div>
-        <select value={this.state.shopItemFilter} onChange={this.handleShopItemFilterChange}>
-          <option key='all' value="all">all</option>
-          {rangeFromTo(1, prosperity + 1).map(level => <option key={level} value={level}>{level}</option>)}
-        </select>
-      </div>
-
-      {
-        this.itemsToDisplay().map(category => {
-          return (
-            <div key={category.level}>
-              <h3 key={"h3-" + category.level}>Prosperity {category.level}</h3>
-              {category.items.map(itemToDiv)}
-            </div>
-          )
-        })
-      }
-    </div>;
-  }
-}
-
-interface CancelDialog {
-  cancel: () => void;
-}
-
-interface ImportExportProps extends CancelDialog {
-  stacks: CardStacks,
-  import: (text: string) => void,
-}
-
-interface ImportExportState {
-  stateAsJsonString: string;
-}
-
-class ImportExport extends React.Component<ImportExportProps, ImportExportState> {
-  static supportForCopyToClipboard(): boolean {
-    const maybeClipboard = navigator.clipboard;
-    return maybeClipboard != undefined && (typeof maybeClipboard.readText === "function");
-  }
-
-  static supportForImportFromClipboard(): boolean {
-    const maybeClipboard = navigator.clipboard;
-    return maybeClipboard != undefined && (typeof maybeClipboard.writeText === "function");
-  }
-
-  constructor(props: ImportExportProps) {
-    super(props);
-    this.importClicked = this.importClicked.bind(this);
-    this.stacksToJsonString = this.stacksToJsonString.bind(this);
-    this.copyToClipboard = this.copyToClipboard.bind(this);
-    this.importFromClipboard = this.importFromClipboard.bind(this);
-    this.state = { stateAsJsonString: '' };
-  }
-
-  importClicked() {
-    this.props.import(this.state.stateAsJsonString)
-  }
-
-  stacksToJsonString() {
-    return JSON.stringify(this.props.stacks, null, 2)
-  }
-
-  copyToClipboard() {
-    navigator.clipboard!.writeText(this.stacksToJsonString())
-      .then(() => console.log("copied to clipboard"))
-      .catch((error) => alert(error));
-  }
-
-  importFromClipboard() {
-    navigator.clipboard!.readText().then(this.props.import);
-  }
-
-  onTextChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    this.setState({ stateAsJsonString: e.target.value })
-  };
-
-  render() {
-    return <React.Fragment>
-      <h2 key="h2">Import / Export</h2>
-      <textarea
-        key="textarea" rows={20} cols={20}
-        defaultValue={this.stacksToJsonString()}
-        onChange={this.onTextChange}
-      />
-      <button key="to-clipboard" hidden={!ImportExport.supportForCopyToClipboard()} type="submit" onClick={this.copyToClipboard}>Copy to Clipboard</button>
-      <button key="from-clipboard" hidden={!ImportExport.supportForImportFromClipboard()} type="submit" onClick={this.importFromClipboard}>Import from Clipboard</button>
-      <button key="import" type="submit" onClick={this.importClicked}>{"Import"}</button>
-      <button key="cancel" type="submit" onClick={this.props.cancel}>{"Cancel"}</button>
-    </React.Fragment>;
-  }
-}
-
-function itemsAboveProsperity(title: String, items: Array<number>, prosperity: number) {
-  let maxProsperityItem = itemIdsByProsperityLevel[prosperity].slice(-1)[0];
-  let itemDivs = items.filter(item => item > maxProsperityItem).map(itemToDiv);
-  return <div key={title + "-div"}>
-    {itemDivs.length > 0 && <h2 key="h2">{title}</h2>}
-    {itemDivs}
-  </div>;
-}
 
 export interface CardStacks {
   cityEvents: CardStack;
