+++ src/app/CancelDialog.ts
@@ -1,0 +1,3 @@
+export interface CancelDialog {
+  cancel: () => void;
+}
